Add files via upload
Here is the code for the NeonInspiration Discord Bot.  This bot runs on the NeonPuppy Discord server and its main goal is to cheer up the community.  It will respond to any negative statement, give you quotes whenever you need it, and you can customise what the bot will reply to you when you are sad on Discord.   

Much of the code is based on this video by Freecodecamp teaching how to build a Discord Bot with Python - https://www.youtube.com/watch?v=SPTfmiYiuok